pyproject.toml
[tool.poetry]
name = "kloch"
# keep in sync with __init__.py
version = "0.6.0"
description = "Environment Manager CLI wrapping package managers calls as serialized config file"
authors = ["Liam Collod <monsieurlixm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
python = ">=3.9"
pytest = { version = ">=7.2.0", python = ">=3.7" }
black = { version = "^24.4.2", python = ">=3.9" }
Sphinx = { version = "^7.3.7", python = ">=3.9" }
sphinx-autobuild = { version = "2024.2.4", python = ">=3.9" }
sphinx-exec-code = { version = "0.12", python = ">=3.9" }
sphinx-copybutton = { version = "0.5", python = ">=3.9" }
furo = { version = "^2024.4.27", python = ">=3.9" }
pyinstaller = { version = "^6", python = ">=3.9,<3.13" }
myst_parser = { version = "^3", python = ">=3.9,<3.13" }
nuitka = "^2.2"
# XXX: issue with CI on unix machine
typing-extensions = { version = "<4.8" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
